backend/requirements.txt
fastapi
uvicorn
google-generativeai
faiss-cpu
python-dotenv
